Refatorado MarketApi
-- Refatorado marketApi, tirado as funções de add e remove do módulo da market e colocado no index da marketApi
-- O módulo market agora exporta a classe Market
-- Resolvido um bug no módulo da trade
